Refactor moment offset addition
