+++ packages/utils/src/__tests__/period/period/period.test.js
@@ -6,13 +6,10 @@
 import MockDate from 'mockdate';
 import moment from 'moment';
 
-import { utcMoment } from '../../../datetime';
-
 import {
   dateStringToPeriod,
   findCoarsestPeriodType,
   getCurrentPeriod,
-  getOffsetMoment,
   isCoarserPeriod,
   isFuturePeriod,
   isValidPeriod,
@@ -281,80 +278,4 @@ describe('period utilities', () => {
   });
 
   describe('getPeriodsInRange', testGetPeriodsInRange);
-
-  describe('getOffsetMoment', () => {
-    const currentDateStub = '2020-12-27';
-
-    beforeEach(() => {
-      MockDate.set(currentDateStub);
-    });
-
-    afterEach(() => {
-      MockDate.reset();
-    });
-
-    const testData = [
-      // date
-      ['empty date - uses today', [undefined, {}], '2020-12-27'],
-      // unit, offset
-      ['unit is empty - uses input date', ['2019-06-07', {}], '2019-06-07'],
-      ['unit is invalid - uses input date', ['2019-06-07', { unit: 'invalid' }], '2019-06-07'],
-      ['unit: year,    + offset', ['2019-06-07', { offset: +2, unit: 'YEAR' }], '2021-06-07'],
-      ['unit: year,    - offset', ['2019-06-07', { offset: -2, unit: 'YEAR' }], '2017-06-07'],
-      ['unit: quarter, + offset', ['2019-06-07', { offset: +2, unit: 'QUARTER' }], '2019-12-07'],
-      ['unit: quarter, - offset', ['2019-06-07', { offset: -2, unit: 'QUARTER' }], '2018-12-07'],
-      ['unit: month,   + offset', ['2019-06-07', { offset: +2, unit: 'MONTH' }], '2019-08-07'],
-      ['unit: month,   - offset', ['2019-06-07', { offset: -2, unit: 'MONTH' }], '2019-04-07'],
-      ['unit: week,    + offset', ['2019-06-07', { offset: +2, unit: 'WEEK' }], '2019-06-21'],
-      ['unit: week,    - offset', ['2019-06-07', { offset: -2, unit: 'WEEK' }], '2019-05-24'],
-      ['unit: day,     + offset', ['2019-06-07', { offset: +2, unit: 'DAY' }], '2019-06-09'],
-      ['unit: day,     - offset', ['2019-06-07', { offset: -2, unit: 'DAY' }], '2019-06-05'],
-      [
-        'unit is lowercase - converts to uppercase',
-        ['2019-06-07', { offset: +2, unit: 'year' }],
-        '2021-06-07',
-      ],
-      // modifier, modifierUnit
-      [
-        'modifier is empty - does not modify date',
-        ['2019-06-07', { unit: 'year', modifierUnit: 'month' }],
-        '2019-06-07',
-      ],
-      [
-        'modifier is invalid - does not modify date',
-        ['2019-06-07', { unit: 'year', modifier: 'invalid', modifierUnit: 'month' }],
-        '2019-06-07',
-      ],
-      [
-        'modifier: start_of',
-        ['2019-06-07', { unit: 'year', offset: -2, modifier: 'start_of', modifierUnit: 'year' }],
-        '2017-01-01T00:00:00Z',
-      ],
-      [
-        'modifier: end_of',
-        ['2019-06-07', { unit: 'year', offset: +2, modifier: 'end_of', modifierUnit: 'year' }],
-        '2021-12-31T23:59:59Z',
-      ],
-      [
-        'modifier is valid, modifierUnit is empty - uses unit',
-        ['2019-06-07', { unit: 'year', offset: +2, modifier: 'end_of' }],
-        '2021-12-31T23:59:59Z',
-      ],
-      [
-        'modifier is valid, modifierUnit is not empty - uses modifierUnit',
-        ['2019-06-07', { unit: 'year', offset: +2, modifier: 'end_of', modifierUnit: 'month' }],
-        '2021-06-30T23:59:59Z',
-      ],
-      [
-        'modifier is valid, offset is empty - does not apply offset',
-        ['2019-06-07', { unit: 'year', modifier: 'end_of', modifierUnit: 'year' }],
-        '2019-12-31T23:59:59Z',
-      ],
-    ];
-
-    it.each(testData)('%s', (_, [date, dateOffset], expected) => {
-      const received = getOffsetMoment(date, dateOffset);
-      expect(received.format()).toBe(utcMoment(expected).format());
-    });
-  });
 });
