+++ packages/utils/src/period/index.js
@@ -3,6 +3,7 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
 
+export * from './addMomentOffset';
 export {
   convertDateRangeToPeriods,
   convertDateRangeToPeriodString,
